requirements.txt
streamlit
opencv-python
dlib-bin==19.24.0
numpy
pillow
pandas
